chore: update authentication service and tests for improved rate limiting and coverage reporting
