+++ Sources/MagSafeGuard/Services/AuthenticationService.swift
@@ -89,7 +89,7 @@ public class AuthenticationService: NSObject {
     private var context: LAContext
     
     /// Last successful authentication timestamp
-    private var lastAuthenticationTime: Date?
+    internal var lastAuthenticationTime: Date?
     
     /// Authentication cache duration (5 minutes)
     private let authenticationCacheDuration: TimeInterval = 300
@@ -191,126 +191,10 @@ public class AuthenticationService: NSObject {
         return nil
     }
     
-    /// Configure authentication context based on policy
-    private func configureContext(_ context: LAContext, for policy: AuthenticationPolicy) {
-        // Set timeout to prevent indefinite authentication attempts
-        context.touchIDAuthenticationAllowableReuseDuration = 0
-        
-        // Configure fallback based on policy
-        if policy.contains(.biometricOnly) {
-            context.localizedFallbackTitle = ""
-        } else {
-            context.localizedFallbackTitle = "Enter Password"
-        }
-    }
-    
     /// Perform the actual authentication
     private func performAuthentication(reason: String, policy: AuthenticationPolicy, completion: @escaping (AuthenticationResult) -> Void) {
-        let authContext = LAContext()
-        configureContext(authContext, for: policy)
-        
-        let laPolicy: LAPolicy = policy.contains(.biometricOnly)
-            ? .deviceOwnerAuthenticationWithBiometrics
-            : .deviceOwnerAuthentication
-        
-        var error: NSError?
-        guard authContext.canEvaluatePolicy(laPolicy, error: &error) else {
-            let authError = mapLAError(error)
-            DispatchQueue.main.async {
-                completion(.failure(authError))
-            }
-            return
-        }
-        
-        #if DEBUG
-        print("[AuthenticationService] Authentication requested with reason: \(reason)")
-        #endif
-        
-        // Security Note: We use evaluatePolicy here because it's the official Apple API
-        // for biometric authentication. The Snyk warning about DeviceAuthenticationBypass
-        // is mitigated through our comprehensive security measures:
-        // 1. Rate limiting prevents brute force attempts
-        // 2. Input validation prevents injection attacks
-        // 3. Fresh contexts prevent replay attacks
-        // 4. Production security checks validate authentication state
-        // 5. Attempt tracking monitors for suspicious activity
-        
-        // Additional security: Verify context integrity before evaluation
-        authContext.localizedCancelTitle = "Cancel"
-        
-        // Set interaction not allowed to prevent UI spoofing
-        authContext.interactionNotAllowed = false
-        
-        // Create a completion handler to avoid deep nesting
-        let evaluationCompletion: (Bool, Error?) -> Void = { [weak self] success, error in
-            self?.processAuthenticationResponse(success: success, error: error, context: authContext, completion: completion)
-        }
-        
-        // IMPORTANT: This is a legitimate use of evaluatePolicy for a security application
-        // Alternative approaches (like keychain-based authentication) would not provide
-        // the same level of user verification that biometrics offer
-
-        // deepcode ignore swift/DeviceAuthenticationBypass: This is the official Apple API for biometric authentication. See above., deepcode ignore DeviceAuthenticationBypass: We have implemented comprehensive security measures including rate limiting, input validation, attempt tracking, and production security checks to mitigate any potential bypass attempts. For a security application like MagSafe Guard, biometric authentication is essential.
-        authContext.evaluatePolicy(laPolicy, localizedReason: reason, reply: evaluationCompletion)
-    }
-    
-    /// Process authentication response (separate method to reduce nesting)
-    private func processAuthenticationResponse(success: Bool, error: Error?, context: LAContext, completion: @escaping (AuthenticationResult) -> Void) {
-        handleAuthenticationResult(success: success, error: error, context: context, completion: completion)
-    }
-    
-    /// Handle authentication result
-    private func handleAuthenticationResult(success: Bool, error: Error?, context: LAContext, completion: @escaping (AuthenticationResult) -> Void) {
-        if success {
-            handleAuthenticationSuccess(context: context, completion: completion)
-        } else if let error = error {
-            handleAuthenticationError(error as NSError, completion: completion)
-        } else {
-            recordAuthenticationAttempt(success: false)
-            DispatchQueue.main.async {
-                completion(.failure(AuthenticationError.unknown(NSError(domain: "AuthenticationService", code: -1, userInfo: nil))))
-            }
-        }
-    }
-    
-    /// Handle successful authentication
-    private func handleAuthenticationSuccess(context: LAContext, completion: @escaping (AuthenticationResult) -> Void) {
-        #if !DEBUG
-        // Production security check
-        guard context.evaluatedPolicyDomainState != nil else {
-            recordAuthenticationAttempt(success: false)
-            DispatchQueue.main.async {
-                completion(.failure(AuthenticationError.authenticationFailed))
-            }
-            return
-        }
-        #endif
-        
-        recordAuthenticationAttempt(success: true)
-        lastAuthenticationTime = Date()
-        DispatchQueue.main.async {
-            completion(.success)
-        }
-    }
-    
-    /// Handle authentication error
-    private func handleAuthenticationError(_ error: NSError, completion: @escaping (AuthenticationResult) -> Void) {
-        let authError = mapLAError(error)
-        
-        // Only record failed attempts for actual authentication failures
-        if case .userCancel = authError {
-            // Don't count cancellations as failed attempts
-        } else {
-            recordAuthenticationAttempt(success: false)
-        }
-        
-        DispatchQueue.main.async {
-            if case .userCancel = authError {
-                completion(.cancelled)
-            } else {
-                completion(.failure(authError))
-            }
-        }
+        // Delegate to LAContext-specific implementation
+        performAuthenticationWithLAContext(reason: reason, policy: policy, completion: completion)
     }
     
     /// Invalidate the current authentication context
