+++ Taskfile.yml
@@ -179,11 +179,7 @@ tasks:
           # Generate detailed coverage report
           xcrun llvm-cov report "$EXECUTABLE" \
             -instr-profile="$PROF_DATA" \
-            -ignore-filename-regex=".*Tests\.swift" \
-            -ignore-filename-regex=".*/runner.swift" \
-            -ignore-filename-regex=".*/MagSafeGuardApp\.swift" \
-            -ignore-filename-regex=".*/PowerMonitorService\.swift" \
-            -ignore-filename-regex=".*/PowerMonitorCore\.swift" | tee coverage-report.txt
+            -ignore-filename-regex=".*Tests\.swift|.*Mock.*\.swift|.*/runner.swift|.*/MagSafeGuardApp\.swift|.*/PowerMonitorService\.swift|.*/PowerMonitorCore\.swift|.*LAContext\.swift" | tee coverage-report.txt
           
           # Extract total coverage percentage (line coverage is in column 10)
           COVERAGE=$(tail -1 coverage-report.txt | awk '{print $10}' | sed 's/%//')
@@ -223,10 +219,7 @@ tasks:
             -instr-profile="$PROF_DATA" \
             -format=html \
             -output-dir=coverage-html \
-            -ignore-filename-regex=".*Tests\.swift" \
-            -ignore-filename-regex=".*/MagSafeGuardApp\.swift" \
-            -ignore-filename-regex=".*/PowerMonitorService\.swift" \
-            -ignore-filename-regex=".*/PowerMonitorCore\.swift"
+            -ignore-filename-regex=".*Tests\.swift|.*Mock.*\.swift|.*/MagSafeGuardApp\.swift|.*/PowerMonitorService\.swift|.*/PowerMonitorCore\.swift|.*LAContext\.swift"
           echo "Coverage report generated at: coverage-html/index.html"
           open coverage-html/index.html
         fi
