+++ sonar-project.properties
@@ -20,7 +20,7 @@ sonar.swift.coverage.reportPath=coverage.lcov
 
 # Exclusions
 sonar.exclusions=.build/**,DerivedData/**,**/*.xcodeproj/**,**/*.md,**/.*,**/*.yml,**/*.yaml,**/*.json,**/Package.swift
-sonar.coverage.exclusions=**/*Tests.swift,**/Mock*.swift,**/MagSafeGuardApp.swift,**/PowerMonitorService.swift,**/PowerMonitorCore.swift
+sonar.coverage.exclusions=**/*Tests.swift,**/Mock*.swift,**/MagSafeGuardApp.swift,**/PowerMonitorService.swift,**/PowerMonitorCore.swift,**/*LAContext*.swift
 
 # Duplication Detection
 sonar.cpd.exclusions=**/*Tests.swift,**/Mock*.swift
